Fix CI failure: Move parsing logic to shared utility src/utils/parser.ts
- Moved parseLogs from observability-ui to src/utils/parser.ts
- Use ConductorEvent from src/utils/logger.ts in the new parser
- Update observability-ui to re-export parseLogs and ConductorEvent from root
- Update tests/utils/parser.test.ts to use the new location

Closes #108
